refactor: NFTBrowse into AssetBrowse
