+++ webapp/src/components/AssetBrowse/index.ts
@@ -1,0 +1,2 @@
+import AssetBrowse from './AssetBrowse.container'
+export { AssetBrowse }
