--- webapp/src/components/NFTBrowse/index.ts
@@ -1,2 +1,0 @@
-import NFTBrowse from './NFTBrowse.container'
-export { NFTBrowse }
